Try to resolve the errors thrown by the checker

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -8,9 +8,9 @@
 
 void print_number(int n)
 {
-long m;
+int m;
 int c;
-long num;
+int num;
 
 num = n;
